Rename namespace Smart.Design.Razor.TagHelpers.FormGroups to Smart.Design.Razor.TagHelpers.FormGroupElements
Implementation under this namespace should be form group having only one HTML element.
