+++ src/Smart.Design.Razor/TagHelpers/FormGroupElements/FormGroupInputGroup.cs
@@ -6,7 +6,7 @@ using Smart.Design.Razor.TagHelpers.Constants;
 using Smart.Design.Razor.TagHelpers.Icon;
 using Smart.Design.Razor.TagHelpers.InputGroup;
 
-namespace Smart.Design.Razor.TagHelpers.FormGroups;
+namespace Smart.Design.Razor.TagHelpers.FormGroupElements;
 
 /// <summary>
 /// <see cref="BaseSmartFormGroupTagHelper"/> implementation with a <see cref="FormGroupInputGroup" />.
